Created usecase class for project completion status
