+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/project/ProjectCompletionStatus.kt
@@ -1,0 +1,31 @@
+package org.wycliffeassociates.otter.common.domain.project
+
+import org.wycliffeassociates.otter.common.data.workbook.Chapter
+import org.wycliffeassociates.otter.common.data.workbook.Workbook
+import org.wycliffeassociates.otter.common.domain.narration.ChapterRepresentation
+import javax.inject.Inject
+
+class ProjectCompletionStatus @Inject constructor() {
+
+    fun getChapterNarrationProgress(workbook: Workbook, chapter: Chapter): Double {
+        val chapterRepresentation = ChapterRepresentation(workbook, chapter)
+        chapterRepresentation.loadFromSerializedVerses()
+
+        return chapterRepresentation.getCompletionProgress()
+    }
+
+    fun getChapterTranslationProgress(chapter: Chapter): Double {
+        val chunkCount = chapter.chunkCount.blockingGet()
+
+        if (chunkCount == 0) return 0.0
+
+        val chunkWithAudio = chapter.chunks
+            .take(1)
+            .map {
+                it.count { it.hasSelectedAudio() }
+            }
+            .blockingFirst()
+
+        return chunkWithAudio.toDouble() / chunkCount
+    }
+}
