+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/project/ProjectFilesAccessor.kt
@@ -56,7 +56,6 @@ import org.wycliffeassociates.otter.common.domain.project.TakeCheckingStatusMap
 import org.wycliffeassociates.otter.common.data.workbook.*
 import org.wycliffeassociates.otter.common.data.workbook.Take
 import org.wycliffeassociates.otter.common.domain.content.FileNamer
-import org.wycliffeassociates.otter.common.domain.narration.ChapterRepresentation
 import org.wycliffeassociates.otter.common.domain.resourcecontainer.project.usfm.getText
 import org.wycliffeassociates.usfmtools.USFMParser
 import org.wycliffeassociates.usfmtools.models.markers.CMarker
@@ -133,13 +132,6 @@ class ProjectFilesAccessor(
         val activeVersesFile = projectDir.resolve(activeVersesPath)
 
         return listOf(chapterNarrationFile, activeVersesFile)
-    }
-
-    fun getNarrationProgress(workbook: Workbook, chapter: Chapter): Double {
-        val chapterRepresentation = ChapterRepresentation(workbook, chapter)
-        chapterRepresentation.loadFromSerializedVerses()
-
-        return chapterRepresentation.getCompletionProgress()
     }
 
     fun isInitialized(): Boolean {
