+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/viewmodel/ExportProjectViewModel.kt
@@ -13,6 +13,7 @@ import org.wycliffeassociates.otter.common.data.workbook.Workbook
 import org.wycliffeassociates.otter.jvm.controls.model.ChapterDescriptor
 import org.wycliffeassociates.otter.common.data.workbook.WorkbookDescriptor
 import org.wycliffeassociates.otter.common.domain.narration.NarrationFactory
+import org.wycliffeassociates.otter.common.domain.project.ProjectCompletionStatus
 import org.wycliffeassociates.otter.common.domain.project.exporter.AudioProjectExporter
 import org.wycliffeassociates.otter.common.domain.project.exporter.ExportOptions
 import org.wycliffeassociates.otter.common.domain.project.exporter.ExportResult
@@ -47,6 +48,9 @@ class ExportProjectViewModel : ViewModel() {
     @Inject
     lateinit var narrationFactory: NarrationFactory
 
+    @Inject
+    lateinit var projectCompletionStatus: ProjectCompletionStatus
+
     init {
         (app as IDependencyGraphProvider).dependencyGraph.inject(this)
     }
@@ -66,25 +70,12 @@ class ExportProjectViewModel : ViewModel() {
 
                 workbook.target.chapters
                     .map { chapter ->
-                        val chunkCount = chapter.chunkCount.blockingGet()
-
                         val progress = when {
                             chapter.hasSelectedAudio() -> 1.0
                             hasInProgressNarration(workbook, chapter) -> {
-                                workbook.projectFilesAccessor.getNarrationProgress(workbook, chapter)
+                                projectCompletionStatus.getChapterNarrationProgress(workbook, chapter)
                             }
-                            chunkCount != 0 -> {
-                                // collect chunks from the relay as soon as it starts emitting (blocking)
-                                val chunkWithAudio = chapter.chunks
-                                    .take(1)
-                                    .map {
-                                        it.count { it.hasSelectedAudio() }
-                                    }
-                                    .blockingFirst()
-
-                                chunkWithAudio.toDouble() / chunkCount
-                            }
-                            else -> 0.0
+                            else -> projectCompletionStatus.getChapterTranslationProgress(chapter)
                         }
 
                         ChapterDescriptor(chapter.sort, progress)
